tech?(api): move script to prescription context
