+++ api/tests/prescription/scripts/integration/delete-and-anonymise-previous-participations_test.js
@@ -1,13 +1,13 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-import { DeleteAndAnonymisePreviousCampaignsScript } from '../../../../scripts/prod/delete-and-anonymise-previous-campaigns.js';
+import { DeleteAndAnonymisePreviousCampaignParticipationsScript } from '../../../../src/prescription/scripts/delete-and-anonymise-previous-participations.js';
 import { databaseBuilder, knex } from '../../../test-helper.js';
 
-describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
+describe('DeleteAndAnonymisePreviousCampaignParticipationsScript', function () {
   describe('Options', function () {
     it('has the correct options', function () {
-      const script = new DeleteAndAnonymisePreviousCampaignsScript();
+      const script = new DeleteAndAnonymisePreviousCampaignParticipationsScript();
       const { options } = script.metaInfo;
 
       expect(options.startArchiveDate).to.deep.include({
@@ -52,7 +52,7 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
     };
 
     beforeEach(async function () {
-      script = new DeleteAndAnonymisePreviousCampaignsScript();
+      script = new DeleteAndAnonymisePreviousCampaignParticipationsScript();
       logger = { info: sinon.spy(), error: sinon.spy() };
       sinon.stub(process, 'env').value({ ENGINEERING_USER_ID });
       now = new Date(startArchiveDate);
@@ -116,21 +116,6 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
           deletedBy: userId,
           archivedAt: null,
         });
-        archivedOrganizationAtDate.campaignDeletedAtDate = databaseBuilder.factory.buildCampaign({
-          targetProfileId,
-          name: 'campaignDeletedAtDate name',
-          title: 'campaignDeletedAtDate title',
-          customLandingPageText: 'campaignDeletedAtDate landing',
-          externalIdHelpImageUrl: 'campaignDeletedAtDate help',
-          alternativeTextToExternalIdHelpImage: 'campaignDeletedAtDate alt help',
-          customResultPageText: 'campaignDeletedAtDate custom text',
-          customResultPageButtonText: 'campaignDeletedAtDate custom button',
-          customResultPageButtonUrl: 'campaignDeletedAtDate custom url',
-          organizationId: archivedOrganizationAtDate.organization.id,
-          deletedAt: now,
-          deletedBy: userId,
-          archivedAt: null,
-        });
 
         // 3
         activeOrganization.campaignDeleted = databaseBuilder.factory.buildCampaign({
@@ -167,200 +152,6 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
         await databaseBuilder.commit();
       });
 
-      describe('#archivedOrganizationBeforeDate', function () {
-        it('should not anonymize deleted campaigns of archived organization before date', async function () {
-          // when
-          await script.handle({ options: { startArchiveDate }, logger });
-
-          const campaignDeleted = await knex('campaigns')
-            .select(
-              'name',
-              'title',
-              'customLandingPageText',
-              'externalIdHelpImageUrl',
-              'alternativeTextToExternalIdHelpImage',
-              'customResultPageText',
-              'customResultPageButtonText',
-              'customResultPageButtonUrl',
-              'deletedAt',
-              'deletedBy',
-              'archivedAt',
-            )
-            .where('id', archivedOrganizationBeforeDate.campaignDeleted.id)
-            .first();
-
-          // then
-          expect(campaignDeleted).deep.equal({
-            name: archivedOrganizationBeforeDate.campaignDeleted.name,
-            title: archivedOrganizationBeforeDate.campaignDeleted.title,
-            customLandingPageText: archivedOrganizationBeforeDate.campaignDeleted.customLandingPageText,
-            externalIdHelpImageUrl: archivedOrganizationBeforeDate.campaignDeleted.externalIdHelpImageUrl,
-            alternativeTextToExternalIdHelpImage:
-              archivedOrganizationBeforeDate.campaignDeleted.alternativeTextToExternalIdHelpImage,
-            customResultPageText: archivedOrganizationBeforeDate.campaignDeleted.customResultPageText,
-            customResultPageButtonText: archivedOrganizationBeforeDate.campaignDeleted.customResultPageButtonText,
-            customResultPageButtonUrl: archivedOrganizationBeforeDate.campaignDeleted.customResultPageButtonUrl,
-            deletedAt: archivedOrganizationBeforeDate.campaignDeleted.deletedAt,
-            deletedBy: archivedOrganizationBeforeDate.campaignDeleted.deletedBy,
-            archivedAt: archivedOrganizationBeforeDate.campaignDeleted.archivedAt,
-          });
-        });
-      });
-
-      describe('#archivedOrganizationAtDate', function () {
-        it('should anonymise deleted campaigns before date of archived organization at date', async function () {
-          // when
-          await script.handle({ options: { startArchiveDate }, logger });
-
-          const campaignDeleted = await knex('campaigns')
-            .select(
-              'name',
-              'title',
-              'customLandingPageText',
-              'externalIdHelpImageUrl',
-              'alternativeTextToExternalIdHelpImage',
-              'customResultPageText',
-              'customResultPageButtonText',
-              'customResultPageButtonUrl',
-              'deletedAt',
-              'deletedBy',
-              'archivedAt',
-            )
-            .where('id', archivedOrganizationAtDate.campaignDeleted.id)
-            .first();
-
-          // then
-          expect(campaignDeleted).deep.equal({
-            name: '(anonymized)',
-            title: null,
-            customLandingPageText: null,
-            externalIdHelpImageUrl: null,
-            alternativeTextToExternalIdHelpImage: null,
-            customResultPageText: null,
-            customResultPageButtonText: null,
-            customResultPageButtonUrl: null,
-            deletedAt: archivedOrganizationAtDate.campaignDeleted.deletedAt,
-            deletedBy: archivedOrganizationAtDate.campaignDeleted.deletedBy,
-            archivedAt: null,
-          });
-        });
-
-        it('should not anonymise deleted campaigns at date of archived organization at date', async function () {
-          // when
-          await script.handle({ options: { startArchiveDate }, logger });
-
-          const campaignDeleted = await knex('campaigns')
-            .select(
-              'name',
-              'title',
-              'customLandingPageText',
-              'externalIdHelpImageUrl',
-              'alternativeTextToExternalIdHelpImage',
-              'customResultPageText',
-              'customResultPageButtonText',
-              'customResultPageButtonUrl',
-              'deletedAt',
-              'deletedBy',
-              'archivedAt',
-            )
-            .where('id', archivedOrganizationAtDate.campaignDeletedAtDate.id)
-            .first();
-
-          // then
-          expect(campaignDeleted).deep.equal({
-            name: archivedOrganizationAtDate.campaignDeletedAtDate.name,
-            title: archivedOrganizationAtDate.campaignDeletedAtDate.title,
-            customLandingPageText: archivedOrganizationAtDate.campaignDeletedAtDate.customLandingPageText,
-            externalIdHelpImageUrl: archivedOrganizationAtDate.campaignDeletedAtDate.externalIdHelpImageUrl,
-            alternativeTextToExternalIdHelpImage:
-              archivedOrganizationAtDate.campaignDeletedAtDate.alternativeTextToExternalIdHelpImage,
-            customResultPageText: archivedOrganizationAtDate.campaignDeletedAtDate.customResultPageText,
-            customResultPageButtonText: archivedOrganizationAtDate.campaignDeletedAtDate.customResultPageButtonText,
-            customResultPageButtonUrl: archivedOrganizationAtDate.campaignDeletedAtDate.customResultPageButtonUrl,
-            deletedAt: archivedOrganizationAtDate.campaignDeletedAtDate.deletedAt,
-            deletedBy: archivedOrganizationAtDate.campaignDeletedAtDate.deletedBy,
-            archivedAt: archivedOrganizationAtDate.campaignDeletedAtDate.archivedAt,
-          });
-        });
-      });
-
-      describe('#activeOrganization', function () {
-        it('should not anonymise active campaigns of active organization', async function () {
-          // when
-          await script.handle({ options: { startArchiveDate }, logger });
-
-          const campaignActive = await knex('campaigns')
-            .select(
-              'name',
-              'title',
-              'customLandingPageText',
-              'externalIdHelpImageUrl',
-              'alternativeTextToExternalIdHelpImage',
-              'customResultPageText',
-              'customResultPageButtonText',
-              'customResultPageButtonUrl',
-              'deletedAt',
-              'deletedBy',
-              'archivedAt',
-            )
-            .where('id', activeOrganization.campaignActive.id)
-            .first();
-
-          // then
-          expect(campaignActive).deep.equal({
-            name: activeOrganization.campaignActive.name,
-            title: activeOrganization.campaignActive.title,
-            customLandingPageText: activeOrganization.campaignActive.customLandingPageText,
-            externalIdHelpImageUrl: activeOrganization.campaignActive.externalIdHelpImageUrl,
-            alternativeTextToExternalIdHelpImage:
-              activeOrganization.campaignActive.alternativeTextToExternalIdHelpImage,
-            customResultPageText: activeOrganization.campaignActive.customResultPageText,
-            customResultPageButtonText: activeOrganization.campaignActive.customResultPageButtonText,
-            customResultPageButtonUrl: activeOrganization.campaignActive.customResultPageButtonUrl,
-            deletedAt: activeOrganization.campaignActive.deletedAt,
-            deletedBy: activeOrganization.campaignActive.deletedBy,
-            archivedAt: activeOrganization.campaignActive.archivedAt,
-          });
-        });
-
-        it('should anonymise deleted campaigns of active organization', async function () {
-          // when
-          await script.handle({ options: { startArchiveDate }, logger });
-
-          const campaignDeleted = await knex('campaigns')
-            .select(
-              'name',
-              'title',
-              'customLandingPageText',
-              'externalIdHelpImageUrl',
-              'alternativeTextToExternalIdHelpImage',
-              'customResultPageText',
-              'customResultPageButtonText',
-              'customResultPageButtonUrl',
-              'deletedAt',
-              'deletedBy',
-              'archivedAt',
-            )
-            .where('id', activeOrganization.campaignDeleted.id)
-            .first();
-
-          // then
-          expect(campaignDeleted).deep.equal({
-            name: '(anonymized)',
-            title: null,
-            customLandingPageText: null,
-            externalIdHelpImageUrl: null,
-            alternativeTextToExternalIdHelpImage: null,
-            customResultPageText: null,
-            customResultPageButtonText: null,
-            customResultPageButtonUrl: null,
-            deletedAt: archivedOrganizationAtDate.campaignDeleted.deletedAt,
-            deletedBy: archivedOrganizationAtDate.campaignDeleted.deletedBy,
-            archivedAt: null,
-          });
-        });
-      });
-
       describe('#participations', function () {
         beforeEach(async function () {
           // 1
@@ -391,6 +182,8 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
               firstName: 'johnny',
               lastName: 'be good',
               organizationId: activeOrganization.organization.id,
+              deletedAt: null,
+              deletedBy: null,
             });
 
           await databaseBuilder.commit();
@@ -423,6 +216,7 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
               userId: activeOrganization.activeLearner.userId,
               organizationLearnerId: activeOrganization.activeLearner.id,
               participantExternalId: 'second',
+              isImproved: true,
               campaignId: activeOrganization.campaignActive.id,
               deletedAt: null,
               deletedBy: null,
@@ -442,7 +236,7 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
 
           describe('#participations', function () {
             describe('#archivedOrganizationBeforeDate', function () {
-              it('should not update delete participation of archived organization before date', async function () {
+              it('should not update delete participation of deleted campaign on archived organization before date', async function () {
                 // when
                 await script.handle({
                   options: { startArchiveDate },
@@ -465,7 +259,7 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
             });
 
             describe('#archivedOrganizationAtDate', function () {
-              it('anonymize deleted participation of archived organization at date', async function () {
+              it('should not update deleted participation of deleted campaigns on archived organization at date', async function () {
                 // when
                 await script.handle({
                   options: { startArchiveDate },
@@ -479,8 +273,8 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
 
                 // then
                 expect(deletedParticipation).deep.equal({
-                  userId: null,
-                  participantExternalId: null,
+                  userId: archivedOrganizationAtDate.deletedParticipation.userId,
+                  participantExternalId: archivedOrganizationAtDate.deletedParticipation.participantExternalId,
                   deletedAt: archivedOrganizationAtDate.deletedParticipation.deletedAt,
                   deletedBy: archivedOrganizationAtDate.deletedParticipation.deletedBy,
                 });
@@ -488,7 +282,7 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
             });
 
             describe('#activeOrganization', function () {
-              it('should anonymise deleted participation of active organization', async function () {
+              it('should not anonymise deleted participation of deleted campaign on active organization', async function () {
                 // when
                 await script.handle({
                   options: { startArchiveDate },
@@ -502,14 +296,14 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
 
                 // then
                 expect(deleteParticipation).deep.equal({
-                  userId: null,
-                  participantExternalId: null,
+                  userId: activeOrganization.deletedParticipation.userId,
+                  participantExternalId: activeOrganization.deletedParticipation.participantExternalId,
                   deletedAt: activeOrganization.deletedParticipation.deletedAt,
                   deletedBy: userId,
                 });
               });
 
-              it('should not anonymise or active participation of active organization', async function () {
+              it('should not anonymise of active participation of active organization', async function () {
                 // when
                 await script.handle({
                   options: { startArchiveDate },
@@ -529,17 +323,68 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
                   deletedBy: null,
                 });
               });
+
+              it('should anonymise of deleted participation only of active campaign on active organization', async function () {
+                // when
+                const deletedParticipationActiveCampaign = databaseBuilder.factory.buildCampaignParticipation({
+                  userId: activeOrganization.activeLearner.userId,
+                  organizationLearnerId: activeOrganization.activeLearner.id,
+                  participantExternalId: 'second',
+                  campaignId: activeOrganization.campaignActive.id,
+                  deletedAt: new Date('2024-01-01'),
+                  deletedBy: userId,
+                  isImproved: true,
+                });
+
+                const deletedImprovedParticipationActiveCampaign = databaseBuilder.factory.buildCampaignParticipation({
+                  userId: activeOrganization.activeLearner.userId,
+                  organizationLearnerId: activeOrganization.activeLearner.id,
+                  participantExternalId: 'second',
+                  isImproved: false,
+                  campaignId: activeOrganization.campaignActive.id,
+                  deletedAt: new Date('2024-02-01'),
+                  deletedBy: userId,
+                });
+
+                await databaseBuilder.commit();
+                await script.handle({
+                  options: { startArchiveDate },
+                  logger,
+                });
+
+                const activeParticipation = await knex('campaign-participations')
+                  .select('id', 'userId', 'participantExternalId', 'deletedAt', 'deletedBy')
+                  .whereNull('userId');
+
+                // then
+                expect(activeParticipation).deep.members([
+                  {
+                    id: deletedParticipationActiveCampaign.id,
+                    userId: null,
+                    participantExternalId: null,
+                    deletedAt: deletedParticipationActiveCampaign.deletedAt,
+                    deletedBy: deletedParticipationActiveCampaign.deletedBy,
+                  },
+                  {
+                    id: deletedImprovedParticipationActiveCampaign.id,
+                    userId: null,
+                    participantExternalId: null,
+                    deletedAt: deletedImprovedParticipationActiveCampaign.deletedAt,
+                    deletedBy: deletedImprovedParticipationActiveCampaign.deletedBy,
+                  },
+                ]);
+              });
             });
           });
 
           describe('#assessments', function () {
-            it('should detach assessments and update updatedAt column only on archived organization before date', async function () {
+            it('should detach assessments and update updatedAt column only on deleted participations on active campaign', async function () {
               // given
               // 1
               const assessmentFromDeletedParticipationOnArchivedOrganizationBeforeDate =
                 databaseBuilder.factory.buildAssessment({
                   userId: archivedOrganizationBeforeDate.deletedLearner.userId,
-                  campaignParticipationId: null,
+                  campaignParticipationId: archivedOrganizationBeforeDate.deletedParticipation.id,
                   updatedAt: new Date('2024-01-01'),
                 });
 
@@ -552,9 +397,19 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
                 });
 
               // 3
+              const deletedImprovedParticipationActiveCampaign = databaseBuilder.factory.buildCampaignParticipation({
+                userId: activeOrganization.activeLearner.userId,
+                organizationLearnerId: activeOrganization.activeLearner.id,
+                participantExternalId: 'second',
+                isImproved: false,
+                campaignId: activeOrganization.campaignActive.id,
+                deletedAt: new Date('2024-02-01'),
+                deletedBy: userId,
+              });
+
               const assessmentFromDActiveParticipationOnActiveOrganization = databaseBuilder.factory.buildAssessment({
                 userId: activeOrganization.activeLearner.userId,
-                campaignParticipationId: activeOrganization.activeParticipation.id,
+                campaignParticipationId: deletedImprovedParticipationActiveCampaign.id,
                 updatedAt: new Date('2024-01-01'),
               });
 
@@ -575,12 +430,6 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
                 .whereNull('campaignParticipationId')
                 .pluck('id');
 
-              const previousAnonymuzedAssessments = await knex('assessments')
-                .select('id')
-                .whereNot({ updatedAt: now })
-                .whereNull('campaignParticipationId')
-                .pluck('id');
-
               const activeAssessments = await knex('assessments')
                 .select('id')
                 .whereNot({ updatedAt: now })
@@ -588,19 +437,15 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
                 .pluck('id');
 
               // then
-              expect(activeAssessments).lengthOf(1);
-              expect(activeAssessments).deep.members([assessmentFromDActiveParticipationOnActiveOrganization.id]);
-
-              expect(previousAnonymuzedAssessments).lengthOf(1);
-              expect(previousAnonymuzedAssessments).deep.members([
+              expect(activeAssessments).lengthOf(3);
+              expect(activeAssessments).deep.members([
+                assessmentFromDDeletedParticipationOnActiveOrganization.id,
+                assessmentFromDeletedParticipationOnArchivedOrganizationAtDate.id,
                 assessmentFromDeletedParticipationOnArchivedOrganizationBeforeDate.id,
               ]);
 
-              expect(anonymizedAssessments).lengthOf(2);
-              expect(anonymizedAssessments).deep.members([
-                assessmentFromDeletedParticipationOnArchivedOrganizationAtDate.id,
-                assessmentFromDDeletedParticipationOnActiveOrganization.id,
-              ]);
+              expect(anonymizedAssessments).lengthOf(1);
+              expect(anonymizedAssessments).deep.members([assessmentFromDActiveParticipationOnActiveOrganization.id]);
             });
           });
 
@@ -637,11 +482,28 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
                 campaignParticipationId: activeOrganization.activeParticipation.id,
                 trainingId: training2.id,
                 userId: activeOrganization.activeLearner.userId,
-                updatedAt: new Date('2023-01-01'),
+                updatedAt: new Date('2022-01-01'),
               });
 
               databaseBuilder.factory.buildUserRecommendedTraining({
                 campaignParticipationId: activeOrganization.deletedParticipation.id,
+                trainingId: training2.id,
+                userId: activeOrganization.activeLearner.userId,
+                updatedAt: new Date('2023-01-01'),
+              });
+
+              const deletedImprovedParticipationActiveCampaign = databaseBuilder.factory.buildCampaignParticipation({
+                userId: activeOrganization.activeLearner.userId,
+                organizationLearnerId: activeOrganization.activeLearner.id,
+                participantExternalId: 'second',
+                isImproved: false,
+                campaignId: activeOrganization.campaignActive.id,
+                deletedAt: new Date('2024-02-01'),
+                deletedBy: userId,
+              });
+
+              databaseBuilder.factory.buildUserRecommendedTraining({
+                campaignParticipationId: deletedImprovedParticipationActiveCampaign.id,
                 trainingId: training2.id,
                 userId: activeOrganization.activeLearner.userId,
                 updatedAt: new Date('2023-01-01'),
@@ -662,22 +524,28 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
                 .whereNotNull('campaignParticipationId');
 
               // then
-              expect(anonymizedRecommendedTrainingResults).lengthOf(2);
+              expect(anonymizedRecommendedTrainingResults).lengthOf(1);
               expect(anonymizedRecommendedTrainingResults).deep.members([
                 {
                   campaignParticipationId: null,
-                  userId: archivedOrganizationAtDate.deletedParticipation.userId,
-                },
-                {
-                  campaignParticipationId: null,
-                  userId: activeOrganization.deletedParticipation.userId,
+                  userId: activeOrganization.activeLearner.userId,
                 },
               ]);
-              expect(activeRecommendedTrainings).lengthOf(1);
+              expect(activeRecommendedTrainings).lengthOf(3);
               expect(activeRecommendedTrainings).deep.members([
+                {
+                  campaignParticipationId: archivedOrganizationAtDate.deletedParticipation.id,
+                  userId: archivedOrganizationAtDate.deletedParticipation.userId,
+                  updatedAt: new Date('2024-01-01'),
+                },
                 {
                   campaignParticipationId: activeOrganization.activeParticipation.id,
                   userId: activeOrganization.activeParticipation.userId,
+                  updatedAt: new Date('2022-01-01'),
+                },
+                {
+                  campaignParticipationId: activeOrganization.deletedParticipation.id,
+                  userId: activeOrganization.deletedParticipation.userId,
                   updatedAt: new Date('2023-01-01'),
                 },
               ]);
@@ -734,6 +602,22 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
                 campaignParticipationId: activeOrganization.deletedParticipation.id,
               });
 
+              const deletedImprovedParticipationActiveCampaign = databaseBuilder.factory.buildCampaignParticipation({
+                userId: activeOrganization.activeLearner.userId,
+                organizationLearnerId: activeOrganization.activeLearner.id,
+                participantExternalId: 'second',
+                isImproved: false,
+                campaignId: activeOrganization.campaignActive.id,
+                deletedAt: new Date('2024-02-01'),
+                deletedBy: userId,
+              });
+
+              databaseBuilder.factory.buildBadgeAcquisition({
+                userId: deletedImprovedParticipationActiveCampaign.userId,
+                badgeId: nonCertifiableBadge.id,
+                campaignParticipationId: deletedImprovedParticipationActiveCampaign.id,
+              });
+
               await databaseBuilder.commit();
 
               // when
@@ -744,15 +628,11 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
                 .whereNull('userId');
 
               // then
-              expect(detachBadges).lengthOf(2);
+              expect(detachBadges).lengthOf(1);
               expect(detachBadges).deep.members([
                 {
                   badgeId: nonCertifiableBadge.id,
-                  campaignParticipationId: archivedOrganizationAtDate.deletedParticipation.id,
-                },
-                {
-                  badgeId: nonCertifiableBadge.id,
-                  campaignParticipationId: activeOrganization.deletedParticipation.id,
+                  campaignParticipationId: deletedImprovedParticipationActiveCampaign.id,
                 },
               ]);
             });
@@ -764,10 +644,21 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
                 badgeId: certifiableBadge.id,
                 campaignParticipationId: archivedOrganizationBeforeDate.deletedParticipation.id,
               });
-              databaseBuilder.factory.buildBadgeAcquisition({
+
+              const deletedImprovedParticipationActiveCampaign = databaseBuilder.factory.buildCampaignParticipation({
                 userId: activeOrganization.activeLearner.userId,
+                organizationLearnerId: activeOrganization.activeLearner.id,
+                participantExternalId: 'second',
+                isImproved: false,
+                campaignId: activeOrganization.campaignActive.id,
+                deletedAt: new Date('2024-02-01'),
+                deletedBy: userId,
+              });
+
+              databaseBuilder.factory.buildBadgeAcquisition({
+                userId: deletedImprovedParticipationActiveCampaign.userId,
                 badgeId: certifiableBadge.id,
-                campaignParticipationId: activeOrganization.deletedParticipation.id,
+                campaignParticipationId: deletedImprovedParticipationActiveCampaign.id,
               });
 
               await databaseBuilder.commit();
@@ -791,8 +682,8 @@ describe('DeleteAndAnonymisePreviousCampaignsScript', function () {
                 },
                 {
                   badgeId: certifiableBadge.id,
-                  campaignParticipationId: activeOrganization.deletedParticipation.id,
-                  userId: activeOrganization.activeLearner.userId,
+                  campaignParticipationId: deletedImprovedParticipationActiveCampaign.id,
+                  userId: deletedImprovedParticipationActiveCampaign.userId,
                 },
               ]);
             });
