+++ api/tests/prescription/scripts/integration/insert-missing-pole-emploi-sending-from-date_test.js
@@ -1,8 +1,8 @@
 import dayjs from 'dayjs';
 
-import { insertMissingPoleEmploiSendingFromDate } from '../../../../scripts/prod/insert-missing-pole-emploi-sending-from-date.js';
 import { Tag } from '../../../../src/organizational-entities/domain/models/Tag.js';
 import { PoleEmploiSending } from '../../../../src/prescription/campaign-participation/domain/models/PoleEmploiSending.js';
+import { insertMissingPoleEmploiSendingFromDate } from '../../../../src/prescription/scripts/insert-missing-pole-emploi-sending-from-date.js';
 import { CampaignParticipationStatuses, CampaignTypes } from '../../../../src/prescription/shared/domain/constants.js';
 import { Assessment } from '../../../../src/shared/domain/models/Assessment.js';
 import {
